+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes, updates, and fixes to pod5 will be documented here
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.32]
+
+### Changed
+
+- Order of `pod5 view` is backwards compatible with 0.3.30, new `open_pore_level` field is at the end of the list.
+
 ## [0.3.31]
 
 ### Added